pi/requirements.txt
fastapi[standard]
pydantic
dotenv
openai
RPi.GPIO
